numerous bugfixes

+++ src/game/checkpoint.cpp
@@ -13,7 +13,7 @@
 std::vector<Checkpoint> Checkpoint::sCheckpoints;
 
 
-Checkpoint* Checkpoint::getCheckpoint(int32_t index)
+Checkpoint* Checkpoint::getCheckpoint(uint32_t index)
 {
     Checkpoint* cp = nullptr;
 
@@ -33,7 +33,7 @@ Checkpoint* Checkpoint::getCheckpoint(int32_t index)
 }
 
 
-int32_t Checkpoint::add(TmxObject* tmxObject)
+uint32_t Checkpoint::add(TmxObject* tmxObject)
 {
    Checkpoint cp;
 
@@ -51,7 +51,7 @@ int32_t Checkpoint::add(TmxObject* tmxObject)
       auto it = tmxObject->mProperties->mMap.find("index");
       if (it != tmxObject->mProperties->mMap.end())
       {
-         cp.mIndex = it->second->mValueInt;
+         cp.mIndex = static_cast<uint32_t>(it->second->mValueInt);
       }
    }
